Revert "Add PDF417 and MRZ scanners"
This reverts commit a8e87112d96114b3b2cfb4f280159fea8bdfe4c0.
